WarDeclarator-NukeopsRule
