+++ Content.Server/NukeOps/WarDeclaratorSystem.cs
@@ -22,16 +22,12 @@ namespace Content.Server.NukeOps;
 public sealed class WarDeclaratorSystem : EntitySystem
 {
     [Dependency] private readonly IAdminLogManager _adminLogger = default!;
-    [Dependency] private readonly AlertLevelSystem _alertLevelSystem = default!;
     [Dependency] private readonly IConfigurationManager _cfg = default!;
     [Dependency] private readonly IGameTiming _gameTiming = default!;
     [Dependency] private readonly UserInterfaceSystem _userInterfaceSystem = default!;
     [Dependency] private readonly ChatSystem _chat = default!;
     [Dependency] private readonly PopupSystem _popupSystem = default!;
-    [Dependency] private readonly StationSystem _stationSystem = default!;
     [Dependency] private readonly AccessReaderSystem _accessReaderSystem = default!;
-
-    private bool alertLevelChangedFlag = false;
 
     public override void Initialize()
     {
@@ -39,31 +35,6 @@ public sealed class WarDeclaratorSystem : EntitySystem
 
         SubscribeLocalEvent<WarDeclaratorComponent, ActivatableUIOpenAttemptEvent>(OnAttemptOpenUI);
         SubscribeLocalEvent<WarDeclaratorComponent, WarDeclaratorActivateMessage>(OnActivated);
-    }
-
-    // To change the alert level after declaring war
-    public override void Update(float frameTime)
-    {
-        var query = EntityQueryEnumerator<WarDeclaratorComponent>();
-        while (query.MoveNext(out var uid, out var comp))
-        {
-            if (comp.CurrentStatus == WarConditionStatus.WarReady)
-            {
-
-                if (alertLevelChangedFlag)
-                    return;
-
-                if (comp.SetAlertlevel == null)
-                    continue;
-
-                if (_gameTiming.CurTime < comp.AlertlevelTime)
-                    continue;
-
-                SetAlertlevel(comp.SetAlertlevel);
-
-                alertLevelChangedFlag = true;
-            }
-        }
     }
 
     private void OnMapInit(Entity<WarDeclaratorComponent> ent, ref MapInitEvent args)
@@ -108,29 +79,9 @@ public sealed class WarDeclaratorSystem : EntitySystem
             var title = Loc.GetString(ent.Comp.SenderTitle);
             _chat.DispatchGlobalAnnouncement(ent.Comp.Message, title, true, ent.Comp.Sound, ent.Comp.Color);
             _adminLogger.Add(LogType.Chat, LogImpact.Low, $"{ToPrettyString(args.Actor):player} has declared war with this text: {ent.Comp.Message}");
-
-            // Setting the time to change the alert level
-            if (ent.Comp.SetAlertlevel != null)
-            {
-                ent.Comp.AlertlevelTime = _gameTiming.CurTime + TimeSpan.FromSeconds(ent.Comp.AlertlevelDelay);
-            }
         }
 
         UpdateUI(ent, ev.Status);
-    }
-
-    private void SetAlertlevel(string level)
-    {
-        var query = EntityQueryEnumerator<NukeDiskComponent>();
-        while (query.MoveNext(out var uid, out var comp))
-        {
-            var station = _stationSystem.GetOwningStation(uid);
-
-            if (station == null)
-                return;
-
-            _alertLevelSystem.SetLevel(station.Value, level, true, true, true, true);
-        }
     }
 
     private void UpdateUI(Entity<WarDeclaratorComponent> ent, WarConditionStatus? status = null)
